v2.1.0 bug fixes w.r.t failed vercel build - reduced /og-image by refactoring `config.ts` - removed next-compose-plugins - updated CHANGELOG.md
