+++ package.json
@@ -21,7 +21,6 @@
         "ioredis": "^5.0.4",
         "lqip-modern": "^1.2.0",
         "next": "^12.2.5",
-        "next-compose-plugins": "^2.2.1",
         "next-seo": "^5.1.0",
         "node-html-parser": "^5.3.3",
         "notion-compat": "^6.12.10",
@@ -67,7 +66,7 @@
         "fsevents": "2.1.2"
     },
     "overrides": {
-        "lqip-modern" : {
+        "lqip-modern": {
             "sharp": "$sharp"
         }
     }
